+++ src/main/java/io/github/daegwon/ticketing_system/controller/TicketIssueControllerV2.java
@@ -1,7 +1,7 @@
 package io.github.daegwon.ticketing_system.controller;
 
 import io.github.daegwon.ticketing_system.dto.TicketIssueRequestDto;
-import io.github.daegwon.ticketing_system.service.TicketIssueServiceV2_1;
+import io.github.daegwon.ticketing_system.service.ticket_issue.TicketIssueServiceV2_1;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
